In some case, the integer length prefix is not enough so I rewrite to long length prefix

+++ SuffixArrayRun.java
@@ -31,8 +31,8 @@ public class SuffixArrayRun{
         job.setJarByClass(SuffixArrayRun.class);
      
         job.setMapperClass(BioMapper.class);
-        job.setMapOutputKeyClass(IntWritable.class);
-        //job.setMapOutputKeyClass(LongWritable.class);
+        //job.setMapOutputKeyClass(IntWritable.class);
+        job.setMapOutputKeyClass(LongWritable.class);
         job.setMapOutputValueClass(LongWritable.class);
 
         job.setReducerClass(BioReducer.class);
